Crypto Tracker new TabView design
The Crypto Tracker app in mobile now has a TabView that changes from a ChartView to a DataGridView.
