+++ Samples/CryptoTracker/CryptoTracker/CryptoTracker.csproj
@@ -98,6 +98,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="Telerik.UI.for.Maui" Version="0.4.0" />
+	  <PackageReference Include="Telerik.UI.for.Maui" Version="0.5.0" />
 	</ItemGroup>
 </Project>
