+++ Samples/CryptoTracker/CryptoTracker/Pages/CoinSelectionPage.xaml
@@ -5,5 +5,5 @@
              Title="Crypto Tracker"
              x:Name="mainPage">
 
-    <view:CoinSelectionView CoinSelected="OnCoinSelected"/>
+    <view:CoinSelectionView CoinSelected="OnCoinSelected" Padding="20, 25, 25, 15"/>
 </ContentPage>
