+++ Samples/CryptoTracker/CryptoTracker/Platforms/Windows/Package.appxmanifest
@@ -14,7 +14,7 @@
   <Properties>
     <DisplayName>CryptoTracker</DisplayName>
     <PublisherDisplayName>Microsoft</PublisherDisplayName>
-    <Logo>Assets\appiconStoreLogo.png</Logo>
+    <Logo>appiconStoreLogo.png</Logo>
   </Properties>
 
   <Dependencies>
@@ -34,19 +34,19 @@
         DisplayName="CryptoTracker"
         Description="CryptoTracker"
         BackgroundColor="transparent"
-        Square150x150Logo="Assets\appiconMediumTile.png"
-        Square44x44Logo="Assets\appiconLogo.png">
+        Square150x150Logo="appiconMediumTile.png"
+        Square44x44Logo="appiconLogo.png">
         <uap:DefaultTile
-          Wide310x150Logo="Assets\appiconWideTile.png"
-          Square71x71Logo="Assets\appiconSmallTile.png"
-          Square310x310Logo="Assets\appiconLargeTile.png"
+          Wide310x150Logo="appiconWideTile.png"
+          Square71x71Logo="appiconSmallTile.png"
+          Square310x310Logo="appiconLargeTile.png"
           ShortName="CryptoTracker">
           <uap:ShowNameOnTiles>
             <uap:ShowOn Tile="square150x150Logo"/>
             <uap:ShowOn Tile="wide310x150Logo"/>
           </uap:ShowNameOnTiles>
         </uap:DefaultTile >
-        <uap:SplashScreen Image="Assets\loadingSplashScreen.png" />
+        <uap:SplashScreen Image="loadingSplashScreen.png" />
       </uap:VisualElements>
     </Application>
   </Applications>
