+++ Samples/CryptoTracker/CryptoTracker/Views/CoinInfoDataGridView.xaml
@@ -1,0 +1,39 @@
+﻿<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:telerikInput="clr-namespace:Telerik.XamarinForms.Input;assembly=Telerik.Maui.Controls.Compatibility"
+             xmlns:telerikChart="clr-namespace:Telerik.XamarinForms.Chart;assembly=Telerik.Maui.Controls.Compatibility"
+             xmlns:local="clr-namespace:CryptoTracker.ViewModels"
+             xmlns:telerikMauiControls="clr-namespace:Telerik.Maui.Controls;assembly=Telerik.Maui.Controls"
+             xmlns:telerikDataGrid="clr-namespace:Telerik.XamarinForms.DataGrid;assembly=Telerik.Maui.Controls.Compatibility"
+             xmlns:formatters="clr-namespace:CryptoTracker.Formatters"
+             x:Class="CryptoTracker.Views.CoinInfoDataGridView"
+             x:Name="coinInfoDataGridView">
+
+
+    <Grid RowDefinitions="Auto, *" ColumnDefinitions="*" Style="{StaticResource TabViewItemGridLayouStyle}">
+        <Label Margin="5,15,5,5"
+                       FontSize="24" 
+                       FontAttributes="Bold"
+                       TextColor="Black"
+                       Grid.Row="0"
+                       Text="Historical Data"
+                       IsVisible="{Binding IsDataGridVisible, Mode=OneWay}" />
+
+        <telerikDataGrid:RadDataGrid Grid.Row="1"
+                                             AutoGenerateColumns="False" 
+                                             IsVisible="{Binding IsDataGridVisible, Mode=OneWay}"
+                                             ItemsSource="{Binding DataForDataGrid}"
+                                             GridLinesThickness="1"
+                                             GridLinesVisibility="Horizontal"
+                                             GridLinesColor="{StaticResource DataGridLineColor}">
+            <telerikDataGrid:RadDataGrid.Columns>
+                <telerikDataGrid:DataGridDateColumn HeaderStyle="{StaticResource dataGridHeaderStyle}" CellContentStyle="{StaticResource dataGridColumnStyle}" CellContentFormat="{}{0:MMM d, yyyy}" PropertyName="Date" HeaderText="Date"/>
+                <telerikDataGrid:DataGridNumericalColumn HeaderStyle="{StaticResource dataGridHeaderStyle}" CellContentStyle="{StaticResource dataGridColumnStyle}" CellContentFormat="{}${0:N2}" PropertyName="OpeningPrice" HeaderText="Open"/>
+                <telerikDataGrid:DataGridNumericalColumn HeaderStyle="{StaticResource dataGridHeaderStyle}" CellContentStyle="{StaticResource dataGridColumnStyle}" CellContentFormat="{}${0:N2}" PropertyName="Price24High" HeaderText="High"/>
+                <telerikDataGrid:DataGridNumericalColumn HeaderStyle="{StaticResource dataGridHeaderStyle}" CellContentStyle="{StaticResource dataGridColumnStyle}" CellContentFormat="{}${0:N2}" PropertyName="Price24Low" HeaderText="Low"/>
+            </telerikDataGrid:RadDataGrid.Columns>
+        </telerikDataGrid:RadDataGrid>
+    </Grid>
+
+
+</ContentView>
