+++ Samples/CryptoTracker/CryptoTracker/Views/CoinInfoDataGridView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using CryptoTracker.ViewModels;
+using CryptoTracker.Data;
+using Telerik.XamarinForms.Common;
+using Microsoft.Maui.Controls;
+
+namespace CryptoTracker.Views
+{
+    public partial class CoinInfoDataGridView : ContentView
+    {
+        public CoinInfoDataGridView()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
