+++ Samples/CryptoTracker/CryptoTracker/Views/CoinInfoView.xaml.cs
@@ -19,40 +19,7 @@ namespace CryptoTracker.Views
         {
             var viewModel = (CoinInfoViewModel)this.BindingContext;
             viewModel.InitializeCoinData(coinInfo);
-        }
 
-        public void ChangedTimePeriod(object sender, ValueChangedEventArgs<int> e)
-        {
-            if (this.chartTypesSegmentedControl == null)
-            {
-                return;
-            }
-
-            if (e.NewValue == 0)
-            {
-                this.chartTypesSegmentedControl.SetSegmentEnabled(1, false);
-            }
-            else
-            {
-                this.chartTypesSegmentedControl.SetSegmentEnabled(1, true);
-            }
-        }
-
-        public void ChangedChartType(object sender, ValueChangedEventArgs<int> e)
-        {
-            if (this.timePeriodsSegmentedControl == null)
-            {
-                return;
-            }
-
-            if (e.NewValue == 1)
-            {
-                this.timePeriodsSegmentedControl.SetSegmentEnabled(0, false);
-            }
-            else
-            {
-                this.timePeriodsSegmentedControl.SetSegmentEnabled(0, true);
-            }
         }
 
         private void RadCartesianChart_HandlerChanged(object sender, System.EventArgs e)
@@ -77,4 +44,5 @@ namespace CryptoTracker.Views
 #endif
         }
     }
+
 }
