+++ Samples/CryptoTracker/CryptoTracker/Views/CoinSelectionView.xaml
@@ -8,12 +8,12 @@
              x:Name="mainView"
              BackgroundColor="AliceBlue">
 
-    <GridLayout RowDefinitions="200, *">
-        <telerikMauiControls:RadBorder Style="{StaticResource ListViewBorderWrapStyle}" Padding="10" Margin="20, 25, 25, 10">
-            <GridLayout RowDefinitions="Auto, *">
+    <Grid RowDefinitions="160, *" RowSpacing="20">
+        <telerikMauiControls:RadBorder Style="{StaticResource ListViewBorderWrapStyle}" Padding="10">
+            <Grid RowDefinitions="Auto, *">
                 <Label Text="Trending" Margin="10" Style="{StaticResource TrendingTitleStyle}"/>
 
-                <telerikDataControls:RadListView GridLayout.Row="1"
+                <telerikDataControls:RadListView Grid.Row="1"
                                                  ItemTapped="OnTrendingListViewItemTapped"
                                                  SelectionMode="None"
                                                  x:Name="trendingListView"
@@ -26,16 +26,16 @@
                         <DataTemplate>
                             <telerikListView:ListViewTemplateCell>
                                 <telerikListView:ListViewTemplateCell.View>
-                                    <GridLayout>
+                                    <Grid>
                                         <telerikMauiControls:RadBorder Style="{StaticResource TrendingListViewItemBorderWrapStyle}">
-                                            <GridLayout ColumnDefinitions="Auto, Auto, *, Auto" RowDefinitions="Auto, *" >
+                                            <Grid ColumnDefinitions="Auto, Auto, *, Auto" RowDefinitions="Auto, *" >
                                                 <Label Style="{StaticResource TrendingCoinsCounterStyle}" Text="{Binding Index}"/>
-                                                <Label GridLayout.Column="1" Style="{StaticResource TrendingCoinSymbolStyle}" Text="{Binding Data.Symbol}"/>
-                                                <Label GridLayout.Column="2" Style="{StaticResource TrendingCoinNameStyle}" Text="{Binding Data.Name}"/>
-                                                <Label GridLayout.Column="3" Style="{StaticResource TrendingCoinPercentage}" Text="{Binding Data.ChangeInPricePercentage, StringFormat='+{0:f2}%'}"/>
-                                            </GridLayout>
+                                                <Label Grid.Column="1" Style="{StaticResource TrendingCoinSymbolStyle}" Text="{Binding Data.Symbol}"/>
+                                                <Label Grid.Column="2" Style="{StaticResource TrendingCoinNameStyle}" Text="{Binding Data.Name}"/>
+                                                <Label Grid.Column="3" Style="{StaticResource TrendingCoinPercentage}" Text="{Binding Data.ChangeInPricePercentage, StringFormat='+{0:f2}%'}"/>
+                                            </Grid>
                                         </telerikMauiControls:RadBorder>
-                                    </GridLayout>
+                                    </Grid>
                                 </telerikListView:ListViewTemplateCell.View>
                             </telerikListView:ListViewTemplateCell>
                         </DataTemplate>
@@ -44,10 +44,10 @@
                         <telerikListView:ListViewItemStyle BorderWidth="0"/>
                     </telerikDataControls:RadListView.ItemStyle>
                 </telerikDataControls:RadListView>
-            </GridLayout>
+            </Grid>
         </telerikMauiControls:RadBorder>
 
-        <telerikMauiControls:RadBorder GridLayout.Row="1" Style="{StaticResource ListViewBorderWrapStyle}" Margin="20, 10, 25, 10">
+        <telerikMauiControls:RadBorder Grid.Row="1" Style="{StaticResource ListViewBorderWrapStyle}">
             <telerikDataControls:RadListView ItemTapped="OnListViewItemTapped"
                                              SelectionMode="None"
                                              x:Name="listView"
@@ -60,10 +60,10 @@
                     <DataTemplate>
                         <telerikListView:ListViewTemplateCell>
                             <telerikListView:ListViewTemplateCell.View>
-                                <GridLayout ColumnDefinitions="Auto, *, Auto, Auto">
+                                <Grid ColumnDefinitions="Auto, *, Auto, Auto">
                                     <Label Style="{StaticResource ListViewCoinSymbolStyle}" Text="{Binding Symbol}"/>
-                                    <Label GridLayout.Column="1" Style="{StaticResource ListViewCoinNameStyle}" Text="{Binding Name}"/>
-                                    <VerticalStackLayout Spacing="5" GridLayout.Column="2" VerticalOptions="Center">
+                                    <Label Grid.Column="1" Style="{StaticResource ListViewCoinNameStyle}" Text="{Binding Name}"/>
+                                    <VerticalStackLayout Spacing="5" Grid.Column="2" VerticalOptions="Center">
                                         <Label HorizontalTextAlignment="End" Style="{StaticResource ListViewCoinPriceStyle}" Text="{Binding ClosingPrice, StringFormat='${0:N2}'}"/>
                                         <Label HorizontalTextAlignment="End" FontSize="14" TextColor="{Binding ChangeInPriceAmount, Converter={StaticResource ValueToColorConverter}}">
                                             <Label.Text>
@@ -75,14 +75,14 @@
                                         </Label>
                                     </VerticalStackLayout>
 
-                                    <Image GridLayout.Column="3"
+                                    <Image Grid.Column="3"
                                            Source="arrowright.png"
                                            Margin="20, 12, 20, 12"
                                            HorizontalOptions="Center"
                                            VerticalOptions="Center"
                                            WidthRequest="24"
                                            HeightRequest="24" />
-                                </GridLayout>
+                                </Grid>
                             </telerikListView:ListViewTemplateCell.View>
                         </telerikListView:ListViewTemplateCell>
                     </DataTemplate>
@@ -97,5 +97,5 @@
             </telerikDataControls:RadListView>
         </telerikMauiControls:RadBorder>
 
-    </GridLayout>
+    </Grid>
 </ContentView>
